requirements.txt
customtkinter==5.2.2
django_recaptcha==4.0.0
mysql_connector_repackaged==0.3.1
Pillow==10.3.0
